gtkwave3-gtk3: 'ghwlib' was renamed to 'libghw
