+++ gtkwave3-gtk3/src/libghw.c
