+++ gtkwave3-gtk3/src/tree.h
@@ -116,7 +116,7 @@ int t_which;		/* 'i' for facs[i] table, value of < 0 means not a full signame */
 uint32_t t_stem;	/* source stem (if >0) for Open Hierarchy Source Def,  see stem_struct_t */
 uint32_t t_istem;	/* source stem (if >0) for Open Hierarchy Source Inst, see stem_struct_t */
 
-unsigned kind : 7; 	/* Kind of the leaf: ghwlib reads this as val & 0x7f so only 7 bits needed */
+unsigned kind : 7; 	/* Kind of the leaf: libghw reads this as val & 0x7f so only 7 bits needed */
 unsigned children_in_gui : 1; /* indicates that the child nodes are in the gtk2 tree, but gets borrowed during tree creation for fast judy sort */
 char name[];		/* C99 */
 };
